Start separating validation code from config loading
