+++ flit/build.py
@@ -7,7 +7,9 @@ from pathlib import Path
 import tarfile
 from tempfile import TemporaryDirectory
 from types import SimpleNamespace
+import sys
 
+from .inifile import ConfigError
 from .sdist import SdistBuilder
 from .wheel import wheel_main, make_wheel_in
 
@@ -33,19 +35,22 @@ def main(ini_file: Path, formats=None):
 
     sdist_info = wheel_info = None
 
-    if 'sdist' in formats:
-        sb = SdistBuilder(ini_file)
-        sdist_file = sb.build()
-        sdist_info = SimpleNamespace(builder=sb, file=sdist_file)
-        # When we're building both, build the wheel from the unpacked sdist.
-        # This helps ensure that the sdist contains all the necessary files.
-        if 'wheel' in formats:
-            dist_dir = ini_file.parent / 'dist'
-            with unpacked_tarball(sdist_file) as tmpdir:
-                log.debug('Building wheel from unpacked sdist %s', tmpdir)
-                tmp_ini_file = Path(tmpdir, ini_file.name)
-                wheel_info = make_wheel_in(tmp_ini_file, dist_dir)
-    elif 'wheel' in formats:
-        wheel_info = wheel_main(ini_file)
+    try:
+        if 'sdist' in formats:
+            sb = SdistBuilder(ini_file)
+            sdist_file = sb.build()
+            sdist_info = SimpleNamespace(builder=sb, file=sdist_file)
+            # When we're building both, build the wheel from the unpacked sdist.
+            # This helps ensure that the sdist contains all the necessary files.
+            if 'wheel' in formats:
+                dist_dir = ini_file.parent / 'dist'
+                with unpacked_tarball(sdist_file) as tmpdir:
+                    log.debug('Building wheel from unpacked sdist %s', tmpdir)
+                    tmp_ini_file = Path(tmpdir, ini_file.name)
+                    wheel_info = make_wheel_in(tmp_ini_file, dist_dir)
+        elif 'wheel' in formats:
+            wheel_info = wheel_main(ini_file)
+    except ConfigError as e:
+        sys.exit('Config error: {}'.format(e))
 
     return SimpleNamespace(wheel=wheel_info, sdist=sdist_info)
